Separando responsabilidades en StringCalculator
