+++ diegoguemes/EneroStringCalculator/StringCalculator.cs
@@ -1,53 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 
 namespace EneroStringCalculator
 {
     public class StringCalculator
     {
-        private const int MAX_LIMIT = 1000;
         private NumberSplitter splitter;
+        private NegativesValidator negativesValidator;
+        private GreaterThanFilter filter;
 
-        public StringCalculator(NumberSplitter splitter)
+        public StringCalculator(NumberSplitter splitter, NegativesValidator negativesValidator, GreaterThanFilter filter)
         {
             this.splitter = splitter;
+            this.negativesValidator = negativesValidator;
+            this.filter = filter;
         }
 
         public int Add(string numbers)
         {
-            if (string.IsNullOrEmpty(numbers))
-                return 0;
             var parsedNumbers = splitter.SplitNumbers(numbers);
-            ValidateNumbers(parsedNumbers);
-            parsedNumbers = FilterNumbersGreaterThan(MAX_LIMIT, parsedNumbers);
+            negativesValidator.ValidateNumbers(parsedNumbers);
+            parsedNumbers = filter.Filter(parsedNumbers);
             return parsedNumbers.Sum();
         }
-
-        private void ValidateNumbers(IEnumerable<int> numbers)
-        {
-            var negatives = GetNegatives(numbers);
-            if (negatives.Count() > 0)
-            {
-                throw new ArgumentException(string.Format("Negatives not allowed: {0}",
-                                                          string.Join(",", negatives)));
-            }
-        }
-
-        private IEnumerable<int> GetNegatives(IEnumerable<int> numbers)
-        {
-            return numbers.Where(IsNegative);
-        }
-
-        private bool IsNegative(int number)
-        {
-            return number < 0;
-        }
-
-        private IEnumerable<int> FilterNumbersGreaterThan(int maxLimit, IEnumerable<int> numbers)
-        {
-            return numbers.Where(n => n <= maxLimit);
-        }
-
     }
 }
